+++ __init__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-from .base import VERSION
+from .common import VERSION
 
 # Export version
 __version__ = '.'.join('%d' % v for v in VERSION[0:3])